Flat branch update. Collection of fixes and small improvements, required for futher module.
- Fixed Makefiles to correct clean/create modules
- Make menu title possible to have non langid values
- Make load_from_file() much more universal
- variable names enhahcements

+++ branches/reyalp-flt/core/gui_menu.h
@@ -49,7 +49,7 @@ typedef struct {
 
 typedef struct {
     char                symbol;
-    short               title;
+    int                 title;
     void                (*on_change)(unsigned int item);
     const CMenuItem     *menu;
 } CMenu;
